updated search page
