+++ Client/src/pages/Search.js
@@ -1,1 +1,100 @@
 //placeholder
+import React, { useState } from "react";
+import { searchNews, getHeadlines } from "../utils/API";
+import { format_date, timeSince } from "../utils/helpers";
+
+const Search = () => {
+  
+  // State objects for the news articles and search string
+  const [news, updateNews] = useState([]);
+  const [searchString, setSearchString] = useState("");
+
+  // Gets the top headlines from the RESTful API
+  const topHeadlines = async (event) => {
+    event.preventDefault();
+
+    try {
+      const searchOptions = {};
+      const response = await getHeadlines(searchOptions);
+      if (!response.ok) {
+        throw new Error("Something went wrong!");
+      }
+      const { articles } = await response.json();
+      updateNews(articles);
+    } catch (err) {
+      console.log(err);
+    }
+
+    setSearchString("");
+  };
+
+  // Gets the returned articles based on what's in the searchstirng
+  const searchForNews = async (event) => {
+    event.preventDefault();
+
+    if (!searchString) {
+      return false;
+    }
+
+    try {
+      const searchOptions = { searchString };
+      const response = await searchNews(searchOptions);
+      if (!response.ok) {
+        throw new Error("Something went wrong!");
+      }
+      const { articles } = await response.json();
+      updateNews(articles);
+      console.log(articles);
+    } catch (err) {
+      console.log(err);
+    }
+
+    setSearchString("");
+  };
+
+  return (
+    <div>
+      <form>
+        <h2>Search for news</h2>
+        <input
+          type="text"
+          name="searchText"
+          id="seartchText"
+          value={searchString}
+          onChange={(e) => setSearchString(e.target.value)}
+          placeholder="Search for news"
+        />
+        <button onClick={searchForNews}>Search</button>
+        <button onClick={topHeadlines}>Headlines</button>
+      </form>
+
+      {news.map((item, i) => {
+        return (
+          <div key={i}>
+            <div className="headline">
+              <img
+                src={item.urlToImage}
+                alt=""
+                width="150px"
+                className="headline-img"
+              />
+              <div>
+                <a href={item.url}>
+                  <h3>{item.title}</h3>
+                </a>
+                <p>{item.description}</p>
+                <div>
+                  {item.source.name || ""}|{format_date(item.publishedAt)}{" "}
+                  {timeSince(item.publishedAt)}{" "}
+                </div>
+                <p></p>
+              </div>
+            </div>
+          </div>
+        );
+      })}
+    </div>
+  );
+};
+
+export default Search;
